Copy over changes from https://github.com/OpenSourceEconomics/epp_topics/pull/39. Fixes #6.
