+++ .gitignore
@@ -33,6 +33,7 @@ dist
 index.html
 .remote-assets
 components.d.ts
+.slides_pdf
 
 # nonsense
 desktop.ini
