+++ src/ada_topics/config.py
@@ -18,6 +18,7 @@ CHAPTER_NAMES = [
 
 
 SRC = Path(__file__).parent.resolve()
+SLIDES_PDF_DIR = SRC.parent.parent / ".slides_pdf"
 SITE_SOURCE_DIR = SRC.parent.parent / "site_source"
 DATA_DIR = SRC.parent.parent / "piaac"
 SITE_DIR = Path("/home/hmg/admin/website/") / "_static" / "applied-data-analytics"
